+++ docs/apidocs/options.rst
@@ -1,0 +1,4 @@
+.. automodule:: qiskit_ibm_runtime.options
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
