+++ docs/apidocs/runtime_service.rst
@@ -1,0 +1,4 @@
+.. automodule:: qiskit_ibm_runtime
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
