+++ qiskit_ibm_runtime/__init__.py
@@ -191,7 +191,6 @@ Classes
    Estimator
    Sampler
    Session
-   Options
    IBMBackend
    RuntimeJob
    RuntimeProgram
