+++ requirements-dev.txt
@@ -14,7 +14,7 @@ reno>=2.11.0
 scipy>=1.0
 nbformat>=4.4.0
 nbconvert>=5.3.1
-qiskit-aer
+qiskit-aer>=0.11
 websockets>=8
 scikit-quant;platform_system != 'Windows'
 black==22.3.0
